WIP: tie in delete with front-end. refactor and clean dirty code.

+++ shelved/shelvedApp/views.py
@@ -181,25 +181,26 @@ def addMovie(request):
 #@csrf_protect
 # @login_required(login_url='/login')
 def delete(request):
-    requestbody = json.loads(request.body)
+    if request.method == 'DELETE':
+        requestbody = json.loads(request.body)
 
-    # for elem in requestbody:
-    #     print ('elem is ', elem, 'val is :',requestbody[elem])
-    """current user is always empty, no way to find the current user
+        # for elem in requestbody:
+        #     print ('elem is ', elem, 'val is :',requestbody[elem])
+        """current user is always empty, no way to find the current user
 
-    current_user = request.user
-    import pdb; pdb.set_trace()
-    print('User ID is : ' + current_user.id)
-    """
+        current_user = request.user
+        import pdb; pdb.set_trace()
+        print('User ID is : ' + current_user.id)
+        """
 
-    item_type = requestbody['item_type']
-    unique_id = requestbody['unique_id']
-    count = deleteItem(item_type, unique_id)
+        item_type = requestbody['item_type']
+        unique_id = requestbody['unique_id']
+        count = deleteItem(item_type, unique_id)
 
-    data = {}
-    data['deleted_count'] = count
-    json_data = json.dumps(data)
-    return HttpResponse(json_data, status=200);
+        data = {}
+        data['deleted_count'] = count
+        json_data = json.dumps(data)
+        return HttpResponse(json_data, status=200);
 
 
 class movies(APIView):
